Fix postgres failing tests in CI
Turns out that testing for the edit page urls in the e2e tests is
unreliable as the id's aren't guaranteed. The important test is the
response error and draft page messages are found. For good measure we'll
include a test for the first and almost last lines of the output and assert the line count is at least 30.